Conflicts:
	webforms-core/src/main/java/com/biit/webforms/flow/FlowCleaner.java
	webforms-gui/src/main/java/com/biit/webforms/gui/webpages/FlowEditor.java
	webforms-gui/src/main/java/com/biit/webforms/gui/webpages/designer/PropertiesQuestion.java
	webforms-integration-test/src/test/config/testng.xml
	webforms-persistence/src/test/resources/database.properties

+++ webforms-gui/src/main/java/com/biit/webforms/gui/components/ZoomChangedListener.java
@@ -2,6 +2,6 @@ package com.biit.webforms.gui.components;
 
 public interface ZoomChangedListener {
 
-	void zoomChanged(float zoom);
+	void zoomChanged(double zoom);
 
 }
